+++ src/JitterDemo/Playground.cs
@@ -92,7 +92,7 @@ public partial class Playground : RenderWindow
         {
             RigidBody body = World.CreateRigidBody();
             floorShape = new BoxShape(200, 200, 200);
-            body.Position = new JVector(0, -100, 0f);
+            body.Position = new JVector(0, -100, 0);
             body.MotionType = MotionType.Static;
             body.AddShape(floorShape);
         }
@@ -100,7 +100,7 @@ public partial class Playground : RenderWindow
         world.DynamicTree.Filter = World.DefaultDynamicTreeFilter;
         world.BroadPhaseFilter = null;
         world.NarrowPhaseFilter = new TriangleEdgeCollisionFilter();
-        world.Gravity = new JVector(0, -9.81f, 0);
+        world.Gravity = new JVector(0, -9.81d, 0);
         world.SubstepCount = 1;
         world.SolverIterations = (8, 4);
     }
@@ -135,7 +135,7 @@ public partial class Playground : RenderWindow
     public override void Draw()
     {
         // if (Keyboard.KeyPressBegin(Keyboard.Key.P))
-        world.Step(1.0f / 100.0f, multiThread);
+        world.Step(1.0d / 100.0d, multiThread);
 
         UpdateDisplayText();
         LayoutGui();
@@ -155,26 +155,26 @@ public partial class Playground : RenderWindow
             switch (shape)
             {
                 case BoxShape s:
-                    ms = MatrixHelper.CreateScale(s.Size.X, s.Size.Y, s.Size.Z);
+                    ms = MatrixHelper.CreateScale((float)s.Size.X, (float)s.Size.Y, (float)s.Size.Z);
                     boxDrawer.PushMatrix(mat * ms, color);
                     break;
                 case SphereShape s:
-                    ms = MatrixHelper.CreateScale(s.Radius * 2);
+                    ms = MatrixHelper.CreateScale((float)s.Radius * 2);
                     sphereDrawer.PushMatrix(mat * ms, color);
                     break;
                 case CylinderShape s:
-                    ms = MatrixHelper.CreateScale(s.Radius, s.Height, s.Radius);
+                    ms = MatrixHelper.CreateScale((float)s.Radius, (float)s.Height, (float)s.Radius);
                     cylinderDrawer.PushMatrix(mat * ms, color);
                     break;
                 case CapsuleShape s:
-                    ms = MatrixHelper.CreateScale(s.Radius, s.Length, s.Radius);
+                    ms = MatrixHelper.CreateScale((float)s.Radius, (float)s.Length, (float)s.Radius);
                     cylinderDrawer.PushMatrix(mat * ms, color);
-                    ms = MatrixHelper.CreateTranslation(0, 0.5f * s.Length, 0) * MatrixHelper.CreateScale(s.Radius * 2);
+                    ms = MatrixHelper.CreateTranslation(0, (float)(0.5d * s.Length), 0) * MatrixHelper.CreateScale((float)(s.Radius * 2));
                     halfSphereDrawer.PushMatrix(mat * ms, color);
                     halfSphereDrawer.PushMatrix(mat * MatrixHelper.CreateRotationX(MathF.PI) * ms, color);
                     break;
                 case ConeShape s:
-                    ms = MatrixHelper.CreateScale(s.Radius * 2, s.Height, s.Radius * 2);
+                    ms = MatrixHelper.CreateScale((float)(s.Radius * 2), (float)s.Height, (float)(s.Radius * 2));
                     coneDrawer.PushMatrix(mat * ms, color);
                     break;
             }
